requirements.txt
django==1.9.5
MySQL-python
djangorestframework
python-social-auth
python-twitter